fix some boostio build

+++ ubsio-boostio/src/net/net_common.h
@@ -114,7 +114,7 @@ struct NetOptions {
     // Net trans configs
     bool isDevicetrans = false;                          /* net trans service switch */
     bool isSender = false;                               /* net trans service role */
-    uint16_t transDeviceId = 0;                          /* net trans device id */
+    int32_t transDeviceId = 0;                          /* net trans device id */
     std::string deviceTransType = "NULL";                /* net trans type: device_sdma */
     std::string transStoreUrl = "tcp://127.0.0.1:9798";  /* net trans store url */
     uint64_t transMemSize = 1024 * 1024 * 1024;;                        /* net trans memory size */
